% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_forc.R
\name{conditional_forc}
\alias{conditional_forc}
\title{Linear model forecast conditioned on an input forecast}
\usage{
conditional_forc(lm_call, time_vec, ...)
}
\arguments{
\item{lm_call}{Linear model call of the class lm.}

\item{time_vec}{Vector of any class that is equal in length to the data
in \code{lm_call}.}

\item{...}{One or more forecasts of class Forecast, one forecast for each
covariate in the linear model.}
}
\value{
\code{\link{Forecast}} object that contains the conditional forecast.
}
\description{
\code{conditional_forc} takes a linear model call, a vector of time data
associated with the linear model, and a forecast for each covariate in the
linear model. The linear model is estimated once over the entire sample
period and the coefficients are multiplied by the forecasts of each
covariate. Returns a forecast conditional on forecasts of each covariate.
Used to create a forecast for the present period or replicate a forecast made
at a specific period in the past.
}
\examples{
\dontrun{

conditional_forc(
  lm_call = lm(y ~ x1 + x2, data),
  time_vec = data$date,
  x1_forecast, x2_forecast
)

}
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
