% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{forc,Forecast-method}
\alias{forc,Forecast-method}
\title{Get the forecast slot of a Forecast object}
\usage{
\S4method{forc}{Forecast}(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
Vector of forecast values stored in the \link{Forecast} object.
}
\description{
\code{forc} takes a \code{\link{Forecast}} object and gets the
forecast vector of the forecast.
}
\examples{
\dontrun{

forc(Forecast)

}

}
