% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{forc<-}
\alias{forc<-}
\title{Set forecast slot of a Forecast object}
\usage{
forc(Forecast) <- value
}
\arguments{
\item{Forecast}{Forecast object.}

\item{value}{Vector of values assigned to the forecast slot of the Forecast.}
}
\value{
\code{\link{Forecast}} object that contains the new forecast vector.
}
\description{
\code{forc} takes a \code{\link{Forecast}} object and sets the forecast
vector of the forecast.
}
\examples{
\dontrun{

forc(Forecast) <- c(2.45, 2.76, 3.31)

}

}
