% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{forc2df}
\alias{forc2df}
\title{Collect a Forecast object to a data frame}
\usage{
forc2df(...)
}
\arguments{
\item{...}{One or multiple forecasts of the class Forecast.}
}
\value{
\code{data.frame} object that contains forecast information.
}
\description{
\code{forc2df} takes one or more objects of the Forecast class and collects
them into a data frame. Returns a data frame with all of the information that
was stored in the Forecast objects. If multiple forecasts are being
collected, all forecasts must have identical future and realized values.
}
\examples{
\dontrun{

forc2df(x1_forecast)

forc2df(x1_forecast, x2_forecast)

}

}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
