% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{future<-,Forecast-method}
\alias{future<-,Forecast-method}
\title{Set future slot of a Forecast object}
\usage{
\S4method{future}{Forecast}(Forecast) <- value
}
\arguments{
\item{Forecast}{Forecast object.}

\item{value}{Vector of values assigned to the future slot of the Forecast.}
}
\value{
\code{\link{Forecast}} object that contains the new future vector.
}
\description{
\code{future} takes a \code{\link{Forecast}} object and sets the future
vector of the forecast.
}
\examples{
\dontrun{

future(Forecast) <- c("2015-03-01", "2015-03-02", "2015-03-03")

}

}
