% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{future}
\alias{future}
\title{Get the future slot of a Forecast object}
\usage{
future(Forecast)
}
\arguments{
\item{Forecast}{object.}
}
\value{
Vector of future values stored in the \link{Forecast} object.
}
\description{
\code{future} takes a \code{\link{Forecast}} object and gets the future
vector of the forecast.
}
\examples{
\dontrun{

future(Forecast)

}

}
