% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{h_ahead,Forecast-method}
\alias{h_ahead,Forecast-method}
\title{Get the h_ahead slot of a h_ahead object}
\usage{
\S4method{h_ahead}{Forecast}(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
Vector of h_ahead values stored in the \link{Forecast} object.
}
\description{
\code{h_ahead} takes a \code{\link{Forecast}} object and gets the h_ahead
vector of the forecast.
}
\examples{
\dontrun{

h_ahead(Forecast)

}

}
