% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{h_ahead<-}
\alias{h_ahead<-}
\title{Set h_ahead slot of a Forecast object}
\usage{
h_ahead(Forecast) <- value
}
\arguments{
\item{Forecast}{Forecast object.}

\item{value}{Vector of values assigned to the h_ahead slot of the Forecast.}
}
\value{
\code{\link{Forecast}} object that contains the new h_ahead vector.
}
\description{
\code{h_ahead} takes a \code{\link{Forecast}} object and sets the h_ahead
vector of the forecast.
}
\examples{
\dontrun{

h_ahead(Forecast) <- c("2015-03-01", "2015-03-02", "2015-03-03")

}

}
