% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_mean_forc.R
\name{historical_mean_forc}
\alias{historical_mean_forc}
\title{Historical mean forecast}
\usage{
historical_mean_forc(
  realized_vec,
  h_ahead,
  estimation_end,
  time_vec = NULL,
  estimation_window = NULL
)
}
\arguments{
\item{realized_vec}{Vector of realized values. This is the series that is
being forecasted.}

\item{h_ahead}{Integer representing the number of periods ahead that is being
forecasted.}

\item{estimation_end}{Value of any class representing when to end the initial
mean estimation period and begin forecasting.}

\item{time_vec}{Vector of any class that is equal in length to the
\code{realized_vec} vector.}

\item{estimation_window}{Integer representing the number of past periods
that the historical mean should be estimated over in each period.}
}
\value{
\code{\link{Forecast}} object that contains the historical mean
forecast.
}
\description{
\code{historical_mean_forc} takes a vector of realized values, an
integer number of periods ahead to forecast, a period to end the initial
mean estimation and begin forecasting, an optional vector of time data
associated with the realized values, and an optional integer number of past
periods to estimate the mean over. The historical mean is originally
calculated with realized values up to \code{estimation_end} minus the number
of periods specified in \code{estimation_window}. If \code{estimation_window}
is left \code{NULL} then the historical mean is calculated with all available
realized values up to \code{estimation_end}. In each period the historical
mean is set as the \code{h_ahead} period ahead forecast. This process is
iteratively repeated for each period after \code{estimation_end} with the
historical mean updating in each period as more information would have
become available to the forecaster. Returns a historical mean forecast where
the \code{h_ahead} period ahead forecast is simply the historical mean or
rolling window mean of the series being forecasted.
}
\examples{
date <- as.Date(c("2010-03-31", "2010-06-30", "2010-09-30", "2010-12-31",
                  "2011-03-31", "2011-06-30", "2011-09-30", "2011-12-31", 
                  "2012-03-31", "2012-06-30"))
y  <- c(1.09, 1.71, 1.09, 2.46, 1.78, 1.35, 2.89, 2.11, 2.97, 0.99)
data <- data.frame(date, y)

historical_mean_forc(
  realized_vec = data$y,
  h_ahead = 2L,
  estimation_end = as.Date("2011-03-31"),
  time_vec = data$date,
  estimation_window = 4L
)
 
historical_mean_forc(
  realized_vec = data$y,
  h_ahead = 4L,
  estimation_end = 4L
)
 
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
