% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_forc.R
\name{is_forc}
\alias{is_forc}
\title{In-sample linear model forecast}
\usage{
is_forc(lm_call, time_vec = NULL)
}
\arguments{
\item{lm_call}{Linear model call of the class lm.}

\item{time_vec}{Vector of any class that is equal in length to the data
in \code{lm_call}.}
}
\value{
\code{\link{Forecast}} object that contains the in-sample forecast.
}
\description{
\code{is_forc} takes a linear model call and an optional vector of time
data associated with the linear model. The linear model is estimated once
over the entire sample period and the coefficients are multiplied by the
realized values in each period of the sample. Returns an in-sample forecast
conditional on realized values.
}
\examples{
\dontrun{

is_forc(
  lm_call = lm(y ~ x1 + x2, data),
  time_vec = data$date
)

is_forc(
  lm_call = lm(y ~ x1 + x2, data)
)
}
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
