% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{mse}
\alias{mse}
\title{Calculate MSE of a Forecast object}
\usage{
mse(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
MSE value.
}
\description{
\code{mse} takes a \code{\link{Forecast}} object and returns the MSE of the
forecast. MSE is calculated as:
\code{1/length(forecast) * sum((realized - forecast)^2)}
}
\examples{
\dontrun{

mse(Forecast)

}

}
