% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_weighted_forc.R
\name{mse_weighted_forc}
\alias{mse_weighted_forc}
\title{MSE or RMSE weighted forecast}
\usage{
mse_weighted_forc(..., eval_window, errors = "mse", return_weights = FALSE)
}
\arguments{
\item{...}{Two or more forecasts of class Forecast.}

\item{eval_window}{Integer representing the window over which forecast
accuracy is evaluated. Forecasts are weighted based on their accuracy over
the past \code{eval_window} number of periods.}

\item{errors}{Character, either "mse" or "rmse". Selects whether forecast
accuracy is evaluated using mean squared errors or root mean squared
errors.}

\item{return_weights}{Boolean, selects whether the weights used to weight
forecasts in each period are returned. If TRUE, a data frame of weights is returned to the
Global Environment.}
}
\value{
\code{\link{Forecast}} object that contains the weighted forecast.
}
\description{
\code{mse_weighted_forc} takes two or more forecasts, an evaluation window,
and an error function. For each forecast period, the error function is used
to calculate forecast accuracy over the past \code{eval_window} number of
periods. The forecast accuracy of each forecast is used to weight forecasts
based on performance. Returns a weighted forecast. Optionally returns the set
of weights used to weight forecasts in each period.
}
\details{
Forecasts are weighted in each period with the following function. The error
function used is MSE or RMSE depending on user selection. This example shows
MSE errors.
\deqn{weight = (1 / MSE(forecast)) / (1 / sum(MSE(forecasts)))}
}
\examples{
\dontrun{

mse_weighted_forc(
  x1_forecast, x2_forecast,
  eval_window = 4L,
  errors = "mse",
  return_weights = TRUE
)
}
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
