% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_realized_forc.R
\name{oos_realized_forc}
\alias{oos_realized_forc}
\title{Out-of-sample linear model forecast conditioned on realized values}
\usage{
oos_realized_forc(lm_call, h_ahead, estimation_end, time_vec = NULL)
}
\arguments{
\item{lm_call}{Linear model call of the class lm.}

\item{h_ahead}{Integer representing the number of periods ahead that is being
forecasted.}

\item{estimation_end}{Value of any class representing when to end the initial
coefficient estimation period and begin forecasting.}

\item{time_vec}{Vector of any class that is equal in length to the data
in \code{lm_call}.}
}
\value{
\code{\link{Forecast}} object that contains the out-of-sample
forecast.
}
\description{
\code{oos_realized_forc} takes a linear model call, an integer number of
periods ahead to forecast, a period to end the initial coefficient estimation
and begin forecasting, and an optional vector of time data associated with
the linear model. The linear model is originally estimated with data up to
\code{estimation_end}. Coefficients are multiplied by realized values of the
covariates \code{h_ahead} periods ahead to create an \code{h_ahead} period
ahead forecast. This process is iteratively repeated for each period after
\code{estimation_end}  with coefficients updating in each period. Returns an
out-of-sample forecast conditional on realized values that \strong{would not}
have been available at the forecast origin. Tests the out-of-sample
performance of a linear model had it been conditioned on perfect information.
}
\examples{
\dontrun{

oos_realized_forc(
  lm_call = lm(y ~ x1 + x2, data),
  h_ahead = 4L,
  estimation_end = as.Date("2010-01-01"),
  time_vec = data$date
)

oos_realized_forc(
  lm_call = lm(y ~ x1 + x2, data),
  h_ahead = 4L,
  estimation_end = 130L
)

}
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
