% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{origin<-,Forecast-method}
\alias{origin<-,Forecast-method}
\title{Set origin slot of a Forecast object}
\usage{
\S4method{origin}{Forecast}(Forecast) <- value
}
\arguments{
\item{Forecast}{Forecast object.}

\item{value}{Vector of values assigned to the origin slot of the Forecast.}
}
\value{
\code{\link{Forecast}} object that contains the new origin vector.
}
\description{
\code{origin} takes a \code{\link{Forecast}} object and sets the origin
vector of the forecast.
}
\examples{
\dontrun{

origin(Forecast) <- c("2015-01-01", "2015-01-02", "2015-01-03")

}

}
