% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{realized,Forecast-method}
\alias{realized,Forecast-method}
\title{Get the realized slot of a realized object}
\usage{
\S4method{realized}{Forecast}(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
Vector of realized values stored in the \link{Forecast} object.
}
\description{
\code{realized} takes a \code{\link{Forecast}} object and gets the realized
vector of the forecast.
}
\examples{
\dontrun{

realized(Forecast)

}

}
