% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{show,Forecast-method}
\alias{show,Forecast-method}
\title{Print Forecast object to console.}
\usage{
\S4method{show}{Forecast}(object)
}
\arguments{
\item{object}{Forecast object.}
}
\value{
Printed \code{\link{Forecast}} object.
}
\description{
\code{show} takes a \code{\link{Forecast}} object and prints it to console.
}
\examples{
\dontrun{

print(Forecast)

}

}
