% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{str,Forecast-method}
\alias{str,Forecast-method}
\title{Display internal structure structure of Forecast object to the console.}
\usage{
\S4method{str}{Forecast}(object)
}
\arguments{
\item{object}{Forecast object.}
}
\value{
Structure of \code{\link{Forecast}} object.
}
\description{
\code{str} takes a \code{\link{Forecast}} object and prints its internal
structure to the console.
}
\examples{

my_forecast <- Forecast(
  origin   = as.Date(c("2010-03-31", "2010-06-30", "2010-09-30", "2010-12-31")),
  future   = as.Date(c("2011-03-31", "2011-06-30", "2011-09-30", "2011-12-31")),
  forecast = c(4.21, 4.27, 5.32, 5.11),
  realized = c(4.40, 4.45, 4.87, 4.77),
  h_ahead  = 4L
)

str(my_forecast)

}
