% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{[,Forecast-method}
\alias{[,Forecast-method}
\title{Subset Forecast object.}
\usage{
\S4method{[}{Forecast}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{ANY}

\item{i}{ANY}

\item{j}{ANY}

\item{...}{ANY}

\item{drop}{ANY}

\item{Forecast}{Forecast object.}
}
\value{
Subsetted \code{\link{Forecast}} object.
}
\description{
\code{[]} takes a \code{\link{Forecast}} object and subsets it.
}
