% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{collect}
\alias{collect}
\title{Collect a Forecast object to a data frame}
\usage{
collect(...)
}
\arguments{
\item{...}{One or multiple forecasts of the class Forecast.}
}
\value{
\code{data.frame} object that contains forecast information.
}
\description{
\code{collect} takes one or more objects of the Forecast class and collects
them into a data frame. Returns a data frame with all of the information that
was stored in the Forecast objects. If multiple forecasts are being
collected, all forecasts must have identical future and realized values.
}
\examples{
\dontrun{

collect(x1_forecast)

collect(x1_forecast, x2_forecast)

}

}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
