% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_vintage_forc.R
\name{oos_vintage_forc}
\alias{oos_vintage_forc}
\title{Out-of-sample linear model forecast conditioned on vintage forecasts}
\usage{
oos_vintage_forc(lm_call, time_vec, ...)
}
\arguments{
\item{lm_call}{Linear model call of the class lm.}

\item{time_vec}{Vector of any class that is equal in length to the data
in \code{lm_call}.}

\item{...}{Set of forecasts of class Forecast, one forecast for each
covariate in the linear model.}
}
\value{
\code{\link{Forecast}} object that contains the out-of-sample
forecast.
}
\description{
\code{oos_vintage_forc} takes a linear model call, a vector of time data
associated with the linear model, and a forecast for each covariate in the
linear model. For each period in the vintage forecasts, coefficients are
updated based on information that would have been available at the forecast
origin. Coefficients are then multiplied by vintage forecast values. Returns
an out-of-sample forecast conditional on vintage forecasts that
\strong{would} have been available at the forecast origin. Replicates the
forecasts that a linear model would have produced.
}
\examples{
\dontrun{

oos_vintage_forc(
  lm_call = lm(y ~ x1 + x2, data),
  time_vec = data$date,
  x1_forecast_vintage, x2_forecast_vintage
)

}
}
\seealso{
For a detailed example see the help vignette:
\code{vignette("lmForc", package = "lmForc")}
}
