% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{origin}
\alias{origin}
\title{Get the origin slot of a Forecast object}
\usage{
origin(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
Vector of origin values stored in the \link{Forecast} object.
}
\description{
\code{origin} takes a \code{\link{Forecast}} object and gets the origin
vector of the forecast.
}
\examples{
\dontrun{

origin(Forecast)

}

}
