% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{realized<-,Forecast-method}
\alias{realized<-,Forecast-method}
\title{Set realized slot of a Forecast object}
\usage{
\S4method{realized}{Forecast}(Forecast) <- value
}
\arguments{
\item{Forecast}{Forecast object.}

\item{value}{Vector of values assigned to the realized slot of the Forecast.}
}
\value{
\code{\link{Forecast}} object that contains the new realized vector.
}
\description{
\code{realized} takes a \code{\link{Forecast}} object and sets the realized
vector of the forecast.
}
\examples{
\dontrun{

realized(Forecast) <- c("2015-03-01", "2015-03-02", "2015-03-03")

}

}
