% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{rmse,Forecast-method}
\alias{rmse,Forecast-method}
\title{Calculate RMSE of a Forecast object}
\usage{
\S4method{rmse}{Forecast}(Forecast)
}
\arguments{
\item{Forecast}{Forecast object.}
}
\value{
RMSE value.
}
\description{
\code{rmse} takes a \code{\link{Forecast}} object and returns the RMSE of the
forecast. RMSE is calculated as: \code{sqrt(mse)}
}
\examples{
\dontrun{

rmse(Forecast)

}

}
