% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{show,Forecast-method}
\alias{show,Forecast-method}
\title{Print Forecast object to console.}
\usage{
\S4method{show}{Forecast}(object)
}
\arguments{
\item{object}{Forecast object.}
}
\value{
Printed \code{\link{Forecast}} object.
}
\description{
\code{show} takes a \code{\link{Forecast}} object and prints it to console.
}
\examples{

my_forecast <- Forecast(
  origin   = as.Date(c("2010-03-31", "2010-06-30", "2010-09-30", "2010-12-31")),
  future   = as.Date(c("2011-03-31", "2011-06-30", "2011-09-30", "2011-12-31")),
  forecast = c(4.21, 4.27, 5.32, 5.11),
  realized = c(4.40, 4.45, 4.87, 4.77),
  h_ahead  = 4L
)

print(my_forecast)

}
